Interpolate ENV configs that look like environment variables
If a swarm's env_yaml contains values that look like
envvars (e.g. $PATH or $ABC), then interpolate it when writing the
the "env.sh" file during the "setup" phase of the deploy.

Note that `vrun setup` runs as "sudo" so any env var that one wants
available inside the container must be whitelisted in /etc/sudoers.